Create OCLC X--ISSN csv forma
This utility will take an input .csv file that contains a field of ISSNs that one wants to find all related ISSNs for and created an output .txt file of them:
Line 18 specifies which field that ISSN is in the input .csv file (one has to edit this to suit ones input file)
Line 21 specifies the fields in the input file that will be included also in the output file (one has to edit this to suit one's input file)
Line 25 appends the issns that OCLC can match the the inputted issn. These will appear to the right of the input file fields one chooses to include in the output file. There can be an varying number of matches for any inputted ISSNs, but the script will simply list them all (no muss no fuss)
line 26 is unnecessary, but shows the script is working as it prints to the command line all ISSN's it is attempting to match.