Create Compare 2 .csv files
A simple utility that finds matched fields from two input .csv  files, and outputs matched data to a third file .txt based.
To make this work, you will need to customize to your needs:
Line 1 name input file 1 (.csv)
Line 4 name input file 2 (.csv)
Line 7 name output file (.txt)
line 21 name fields you want to compare from file 1 (e) and file2 (e2)
In this case we are comparing field positions 0 and 1  from e (file 1) to field positions  4 and 5 in e2 (file2). This multi field comparison 
was required because the two input files both contained two ISSN fields and we would want to account for any ISSN  between any ISSN field.

line 22 requires you to name the fields from files 1 and 2 that you want to include on the output array (called "int'). These could be the matched ISSN, but could also be any other field that appears in the same row as that of a matched ISSN.

line 25 simple re-states which fields to include on the output file, but based on the new array created in line 22.

There is one major trick to making this work properly: If you choose to compare, or display the last field in either of your two input files, you need to add a field (column) of null values to the right of your input files data, so that--in fact--the last field you are choosing to display is never more than the NEXT to last of the available fields your choosing from your input files.